pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "creatorpack"
version = "0.1.0"
description = "Local-first content repurposing toolkit"
readme = "README.md"
authors = [{name = "CreatorPack Team"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
stt = ["faster-whisper"]
diarization = ["pyannote.audio"]

[project.scripts]
creatorpack = "creatorpack.app_cli.main:main"

[tool.pytest.ini_options]
testpaths = ["creatorpack/tests"]
addopts = "-q"
